--- app/helpers/furimas_helper.rb
@@ -1,2 +1,0 @@
-module FurimasHelper
-end
