+++ app/helpers/products_helper.rb
@@ -1,0 +1,2 @@
+module ProductsHelper
+end
