+++ app/models/prefecture.rb
@@ -19,6 +19,6 @@ class Prefecture < ActiveHash::Base
   ]
 
   include ActiveHash::Associations
-  has_many :items
-  has_many :payments
+  has_many :products
+  # has_many :payments
 end
