+++ app/models/shipping_cost.rb
@@ -6,5 +6,5 @@ class ShippingCost < ActiveHash::Base
   ]
 
   include ActiveHash::Associations
-  has_many :items
+  has_many :products
 end
