+++ app/views/products/edit.html.erb
