+++ app/views/products/index.html.erb
@@ -119,7 +119,7 @@
   <%# 商品一覧 %>
   <div class='item-contents'>
     <h2 class='title'>ピックアップカテゴリー</h2>
-    <%= link_to '新規投稿商品', new_furima_path , class: "subtitle" %>
+    <%= link_to '新規投稿商品', new_product_path , class: "subtitle" %>
     <ul class='item-lists'>
 
       <%# 商品データがある場合は、eachメソッドで一覧表示する %>
@@ -137,10 +137,10 @@
         </div>
         <div class='item-info'>
           <h3 class='item-name'>
-            <%= item.name %>
+            <%= product.name %>
           </h3>
           <div class='item-price'>
-            <span><%= item.price %>円<br><%= item.shipping_cost.name %></span>
+            <span><%= product.price %>円<br><%= product.shipping_cost.name %></span>
             <div class='star-btn'>
               <%= image_tag "star.png", class:"star-icon" %>
               <span class='star-count'>0</span>
@@ -170,7 +170,7 @@
     </ul>
   </div>
 </div>
-<%= link_to(new_furima_path, class: 'purchase-btn') do %>
+<%= link_to(new_product_path, class: 'purchase-btn') do %>
   <span class='purchase-btn-text'>出品する</span>
   <%= image_tag 'icon_camera.png' , size: '185x50' ,class: "purchase-btn-icon" %>
 <% end %>
