+++ app/views/products/new.html.erb
