+++ config/routes.rb
@@ -1,7 +1,7 @@
 Rails.application.routes.draw do
   devise_for :users
   # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
-   root to: 'furimas#index'
+   root to: 'products#index'
    resources :products do
   end
 end
