+++ spec/requests/products_request_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe "Products", type: :request do
+
+end
